superio/smsc/mec1308: Fix AddressMax value for SMBX mailbox
The way this was implemented before was causing ACPI failures.  There
was also a basic misunderstanding of what the AddressMax field was used
for.  In this case, because it's a fixed address, it should be the same
as the AddressMin field.

Getting rid of the addition in the field solves the ACPI output problem.

Change-Id: Idec2bf0ed27ae694e98f141087cdf22401937178
Signed-off-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://review.coreboot.org/14343
Tested-by: build bot (Jenkins)
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-by: Idwer Vollering <vidwer@gmail.com>
Reviewed-by: Duncan Laurie <dlaurie@google.com>